change to module with object export
