+++ experts-api/.gitignore
@@ -105,3 +105,7 @@ dist
 
 # LevelDB database dir
 db/
+
+# tests
+test/
+.test.js
