+++ experts-api/package.json
@@ -27,6 +27,7 @@
     "jsonld": "^8.1.1",
     "jsonld-streaming-parser": "^3.2.0",
     "memory-level": "^1.0.0",
+    "node-fetch": "^3.3.1",
     "quadstore": "^11.0.8",
     "quadstore-comunica": "^3.0.0",
     "rdf-data-factory": "^1.1.0",
